remove old version of DataStructuresFeature
